+++ ddtrace/appsec/_iast/_handlers.py
@@ -4,12 +4,15 @@ import functools
 from wrapt import when_imported
 from wrapt import wrap_function_wrapper as _w
 
-from ddtrace.appsec._handlers import log
 from ddtrace.appsec._iast import _is_iast_enabled
-from ddtrace.appsec._iast._iast_request_context import in_iast_context
-from ddtrace.appsec._iast._patch import if_iast_taint_returned_object_for
-from ddtrace.appsec._iast._patch import if_iast_taint_yield_tuple_for
-from ddtrace.contrib import trace_utils
+from ddtrace.appsec._iast._metrics import _set_metric_iast_instrumented_source
+from ddtrace.appsec._iast._patch import _iast_instrument_starlette_request
+from ddtrace.appsec._iast._patch import _iast_instrument_starlette_request_body
+from ddtrace.appsec._iast._patch import _iast_instrument_starlette_url
+from ddtrace.appsec._iast._patch import _patched_dictionary
+from ddtrace.appsec._iast._patch import try_wrap_function_wrapper
+from ddtrace.appsec._iast._taint_utils import taint_structure
+from ddtrace.internal.logger import get_logger
 
 
 MessageMapContainer = None
@@ -17,6 +20,9 @@ try:
     from google._upb._message import MessageMapContainer  # type: ignore[no-redef]
 except ImportError:
     pass
+
+
+log = get_logger(__name__)
 
 
 def _on_set_http_meta_iast(
@@ -42,6 +48,8 @@ def _on_set_http_meta_iast(
 
 
 def _on_request_init(wrapped, instance, args, kwargs):
+    from ddtrace.appsec._iast._iast_request_context import in_iast_context
+
     wrapped(*args, **kwargs)
     if _is_iast_enabled() and in_iast_context():
         try:
@@ -67,9 +75,6 @@ def _on_request_init(wrapped, instance, args, kwargs):
 
 def _on_flask_patch(flask_version):
     if _is_iast_enabled():
-        from ddtrace.appsec._iast._metrics import _set_metric_iast_instrumented_source
-        from ddtrace.appsec._iast._patch import _patched_dictionary
-        from ddtrace.appsec._iast._patch import try_wrap_function_wrapper
         from ddtrace.appsec._iast._taint_tracking import OriginType
 
         try_wrap_function_wrapper(
@@ -127,9 +132,10 @@ def _on_flask_patch(flask_version):
 
 
 def _on_wsgi_environ(wrapped, _instance, args, kwargs):
+    from ddtrace.appsec._iast._iast_request_context import in_iast_context
+
     if _is_iast_enabled() and args and in_iast_context():
         from ddtrace.appsec._iast._taint_tracking import OriginType
-        from ddtrace.appsec._iast._taint_utils import taint_structure
 
         return wrapped(*((taint_structure(args[0], OriginType.HEADER_NAME, OriginType.HEADER),) + args[1:]), **kwargs)
 
@@ -139,7 +145,6 @@ def _on_wsgi_environ(wrapped, _instance, args, kwargs):
 def _on_django_patch():
     if _is_iast_enabled():
         try:
-            from ddtrace.appsec._iast._metrics import _set_metric_iast_instrumented_source
             from ddtrace.appsec._iast._taint_tracking import OriginType
 
             # we instrument those sources on _on_django_func_wrapped
@@ -153,7 +158,7 @@ def _on_django_patch():
             _set_metric_iast_instrumented_source(OriginType.PARAMETER_NAME)
             _set_metric_iast_instrumented_source(OriginType.BODY)
             when_imported("django.http.request")(
-                lambda m: trace_utils.wrap(
+                lambda m: try_wrap_function_wrapper(
                     m,
                     "QueryDict.__getitem__",
                     functools.partial(if_iast_taint_returned_object_for, OriginType.PARAMETER),
@@ -167,11 +172,11 @@ def _on_django_func_wrapped(fn_args, fn_kwargs, first_arg_expected_type, *_):
     # If IAST is enabled and we're wrapping a Django view call, taint the kwargs (view's
     # path parameters)
     if _is_iast_enabled() and fn_args and isinstance(fn_args[0], first_arg_expected_type):
+        from ddtrace.appsec._iast._iast_request_context import in_iast_context
         from ddtrace.appsec._iast._taint_tracking import OriginType  # noqa: F401
         from ddtrace.appsec._iast._taint_tracking import is_pyobject_tainted
         from ddtrace.appsec._iast._taint_tracking import origin_to_str
         from ddtrace.appsec._iast._taint_tracking import taint_pyobject
-        from ddtrace.appsec._iast._taint_utils import taint_structure
 
         if not in_iast_context():
             return
@@ -240,7 +245,6 @@ def _on_django_func_wrapped(fn_args, fn_kwargs, first_arg_expected_type, *_):
 def _custom_protobuf_getattribute(self, name):
     from ddtrace.appsec._iast._taint_tracking import OriginType
     from ddtrace.appsec._iast._taint_tracking import taint_pyobject
-    from ddtrace.appsec._iast._taint_utils import taint_structure
 
     ret = type(self).__saved_getattr(self, name)
     if isinstance(ret, (str, bytes, bytearray)):
@@ -287,3 +291,106 @@ def _on_grpc_response(message):
     if _is_iast_enabled():
         msg_cls = type(message)
         _patch_protobuf_class(msg_cls)
+
+
+def if_iast_taint_yield_tuple_for(origins, wrapped, instance, args, kwargs):
+    if _is_iast_enabled():
+        from ._iast_request_context import is_iast_request_enabled
+        from ._taint_tracking import taint_pyobject
+
+        if not is_iast_request_enabled():
+            for key, value in wrapped(*args, **kwargs):
+                yield key, value
+        else:
+            for key, value in wrapped(*args, **kwargs):
+                new_key = taint_pyobject(pyobject=key, source_name=key, source_value=key, source_origin=origins[0])
+                new_value = taint_pyobject(
+                    pyobject=value, source_name=key, source_value=value, source_origin=origins[1]
+                )
+                yield new_key, new_value
+
+    else:
+        for key, value in wrapped(*args, **kwargs):
+            yield key, value
+
+
+def if_iast_taint_returned_object_for(origin, wrapped, instance, args, kwargs):
+    value = wrapped(*args, **kwargs)
+    from ._iast_request_context import is_iast_request_enabled
+
+    if _is_iast_enabled() and is_iast_request_enabled():
+        try:
+            from ._taint_tracking import is_pyobject_tainted
+            from ._taint_tracking import taint_pyobject
+
+            if not is_pyobject_tainted(value):
+                name = str(args[0]) if len(args) else "http.request.body"
+                from ddtrace.appsec._iast._taint_tracking import OriginType
+
+                if origin == OriginType.HEADER and name.lower() in ["cookie", "cookies"]:
+                    origin = OriginType.COOKIE
+                return taint_pyobject(pyobject=value, source_name=name, source_value=value, source_origin=origin)
+        except Exception:
+            log.debug("Unexpected exception while tainting pyobject", exc_info=True)
+    return value
+
+
+def _on_iast_fastapi_patch():
+    from ddtrace.appsec._iast._taint_tracking import OriginType
+
+    # Cookies sources
+    try_wrap_function_wrapper(
+        "starlette.requests",
+        "cookie_parser",
+        functools.partial(_patched_dictionary, OriginType.COOKIE_NAME, OriginType.COOKIE),
+    )
+    _set_metric_iast_instrumented_source(OriginType.COOKIE)
+    _set_metric_iast_instrumented_source(OriginType.COOKIE_NAME)
+
+    # Parameter sources
+    try_wrap_function_wrapper(
+        "starlette.datastructures",
+        "QueryParams.__getitem__",
+        functools.partial(if_iast_taint_returned_object_for, OriginType.PARAMETER),
+    )
+    try_wrap_function_wrapper(
+        "starlette.datastructures",
+        "QueryParams.get",
+        functools.partial(if_iast_taint_returned_object_for, OriginType.PARAMETER),
+    )
+    _set_metric_iast_instrumented_source(OriginType.PARAMETER)
+
+    # Header sources
+    try_wrap_function_wrapper(
+        "starlette.datastructures",
+        "Headers.__getitem__",
+        functools.partial(if_iast_taint_returned_object_for, OriginType.HEADER),
+    )
+    try_wrap_function_wrapper(
+        "starlette.datastructures",
+        "Headers.get",
+        functools.partial(if_iast_taint_returned_object_for, OriginType.HEADER),
+    )
+    _set_metric_iast_instrumented_source(OriginType.HEADER)
+
+    # Path source
+    try_wrap_function_wrapper("starlette.datastructures", "URL.__init__", _iast_instrument_starlette_url)
+    _set_metric_iast_instrumented_source(OriginType.PATH)
+
+    # Body source
+    try_wrap_function_wrapper("starlette.requests", "Request.__init__", _iast_instrument_starlette_request)
+    try_wrap_function_wrapper("starlette.requests", "Request.body", _iast_instrument_starlette_request_body)
+    try_wrap_function_wrapper(
+        "starlette.datastructures",
+        "FormData.__getitem__",
+        functools.partial(if_iast_taint_returned_object_for, OriginType.BODY),
+    )
+    try_wrap_function_wrapper(
+        "starlette.datastructures",
+        "FormData.get",
+        functools.partial(if_iast_taint_returned_object_for, OriginType.BODY),
+    )
+    _set_metric_iast_instrumented_source(OriginType.BODY)
+
+    # Instrumented on _iast_starlette_scope_taint
+    _set_metric_iast_instrumented_source(OriginType.PATH_PARAMETER)
