+++ ddtrace/appsec/_iast/_patch.py
@@ -1,4 +1,3 @@
-import functools
 import sys
 from typing import Callable
 from typing import Text
@@ -8,10 +7,7 @@ from wrapt import FunctionWrapper
 from ddtrace.appsec._common_module_patches import wrap_object
 from ddtrace.internal.logger import get_logger
 
-from ._iast_request_context import is_iast_request_enabled
-from ._metrics import _set_metric_iast_instrumented_source
 from ._taint_utils import taint_structure
-from ._utils import _is_iast_enabled
 
 
 log = get_logger(__name__)
@@ -45,111 +41,10 @@ def try_wrap_function_wrapper(module: Text, name: Text, wrapper: Callable):
         log.debug("IAST patching. Module %s.%s not exists", module, name)
 
 
-def if_iast_taint_returned_object_for(origin, wrapped, instance, args, kwargs):
-    value = wrapped(*args, **kwargs)
-
-    if _is_iast_enabled() and is_iast_request_enabled():
-        try:
-            from ._taint_tracking import is_pyobject_tainted
-            from ._taint_tracking import taint_pyobject
-
-            if not is_pyobject_tainted(value):
-                name = str(args[0]) if len(args) else "http.request.body"
-                from ddtrace.appsec._iast._taint_tracking import OriginType
-
-                if origin == OriginType.HEADER and name.lower() in ["cookie", "cookies"]:
-                    origin = OriginType.COOKIE
-                return taint_pyobject(pyobject=value, source_name=name, source_value=value, source_origin=origin)
-        except Exception:
-            log.debug("Unexpected exception while tainting pyobject", exc_info=True)
-    return value
-
-
-def if_iast_taint_yield_tuple_for(origins, wrapped, instance, args, kwargs):
-    if _is_iast_enabled():
-        from ._taint_tracking import taint_pyobject
-
-        if not is_iast_request_enabled():
-            for key, value in wrapped(*args, **kwargs):
-                yield key, value
-        else:
-            for key, value in wrapped(*args, **kwargs):
-                new_key = taint_pyobject(pyobject=key, source_name=key, source_value=key, source_origin=origins[0])
-                new_value = taint_pyobject(
-                    pyobject=value, source_name=key, source_value=value, source_origin=origins[1]
-                )
-                yield new_key, new_value
-
-    else:
-        for key, value in wrapped(*args, **kwargs):
-            yield key, value
-
-
 def _patched_dictionary(origin_key, origin_value, original_func, instance, args, kwargs):
     result = original_func(*args, **kwargs)
 
     return taint_structure(result, origin_key, origin_value, override_pyobject_tainted=True)
-
-
-def _on_iast_fastapi_patch():
-    from ddtrace.appsec._iast._taint_tracking import OriginType
-
-    # Cookies sources
-    try_wrap_function_wrapper(
-        "starlette.requests",
-        "cookie_parser",
-        functools.partial(_patched_dictionary, OriginType.COOKIE_NAME, OriginType.COOKIE),
-    )
-    _set_metric_iast_instrumented_source(OriginType.COOKIE)
-    _set_metric_iast_instrumented_source(OriginType.COOKIE_NAME)
-
-    # Parameter sources
-    try_wrap_function_wrapper(
-        "starlette.datastructures",
-        "QueryParams.__getitem__",
-        functools.partial(if_iast_taint_returned_object_for, OriginType.PARAMETER),
-    )
-    try_wrap_function_wrapper(
-        "starlette.datastructures",
-        "QueryParams.get",
-        functools.partial(if_iast_taint_returned_object_for, OriginType.PARAMETER),
-    )
-    _set_metric_iast_instrumented_source(OriginType.PARAMETER)
-
-    # Header sources
-    try_wrap_function_wrapper(
-        "starlette.datastructures",
-        "Headers.__getitem__",
-        functools.partial(if_iast_taint_returned_object_for, OriginType.HEADER),
-    )
-    try_wrap_function_wrapper(
-        "starlette.datastructures",
-        "Headers.get",
-        functools.partial(if_iast_taint_returned_object_for, OriginType.HEADER),
-    )
-    _set_metric_iast_instrumented_source(OriginType.HEADER)
-
-    # Path source
-    try_wrap_function_wrapper("starlette.datastructures", "URL.__init__", _iast_instrument_starlette_url)
-    _set_metric_iast_instrumented_source(OriginType.PATH)
-
-    # Body source
-    try_wrap_function_wrapper("starlette.requests", "Request.__init__", _iast_instrument_starlette_request)
-    try_wrap_function_wrapper("starlette.requests", "Request.body", _iast_instrument_starlette_request_body)
-    try_wrap_function_wrapper(
-        "starlette.datastructures",
-        "FormData.__getitem__",
-        functools.partial(if_iast_taint_returned_object_for, OriginType.BODY),
-    )
-    try_wrap_function_wrapper(
-        "starlette.datastructures",
-        "FormData.get",
-        functools.partial(if_iast_taint_returned_object_for, OriginType.BODY),
-    )
-    _set_metric_iast_instrumented_source(OriginType.BODY)
-
-    # Instrumented on _iast_starlette_scope_taint
-    _set_metric_iast_instrumented_source(OriginType.PATH_PARAMETER)
 
 
 def _iast_instrument_starlette_url(wrapped, instance, args, kwargs):
